refactor: move render-intent.js into ecs/resources folder (A-12)
