+++ tests/integration/gameplay/d-08-render-dom-system.test.js
@@ -10,7 +10,7 @@
  */
 
 import { describe, expect, it, vi } from 'vitest';
-import { createRenderIntentBuffer } from '../../../src/ecs/render-intent.js';
+import { createRenderIntentBuffer } from '../../../src/ecs/resources/render-intent.js';
 import { createRenderDomSystem } from '../../../src/ecs/systems/render-dom-system.js';
 import { World } from '../../../src/ecs/world/world.js';
